fix: highlight the first word after the cursor too
The original quick-scope vim plugin by Brian Le does actually highlight
the first word away from the cursor, it just doesn't highlight the
current word.

This commit provides that same functionality, or at least brings the
functionality closer to that of the vim plugin, because before this
change, even the first word away from the cursor was not being
highlighted.