fix: inital setting of first-char for forward search, and first-word for backward search, needs to be nil if on a word separator
This is because the list that is looped over for the line, cuts off all
word separators.

This means for forward searching, the first character will actually be
the one to the right of the "current character" (i.e. the character
given by the `following-char` function). And hence this character will
be ignored (that is, not even counted for an occurrence, let alone
highlighted).

For backward searching, this means that the search starts off at the
last letter of the first word (backwards from the cursor). And because
the first word is ignored altogether for highlighting, this word is
therefore never given a highlight, even though it should be, because the
cursor is not actually "on" this word.

Closes #3.
Closes #4.